+++ plugins/webapi/message/plugin.go
@@ -45,7 +45,7 @@ func Plugin() *node.Plugin {
 func GetMessage(c echo.Context) (err error) {
 	messageID, err := messageIDFromContext(c)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, webapi.NewErrorResponse(err))
+		return c.JSON(http.StatusBadRequest, jsonmodels.NewErrorResponse(err))
 	}
 
 	if messagelayer.Tangle().Storage.Message(messageID).Consume(func(message *tangle.Message) {
@@ -54,7 +54,7 @@ func GetMessage(c echo.Context) (err error) {
 		return
 	}
 
-	return c.JSON(http.StatusNotFound, webapi.NewErrorResponse(fmt.Errorf("failed to load Message with %s", messageID)))
+	return c.JSON(http.StatusNotFound, jsonmodels.NewErrorResponse(fmt.Errorf("failed to load Message with %s", messageID)))
 }
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -65,7 +65,7 @@ func GetMessage(c echo.Context) (err error) {
 func GetMessageMetadata(c echo.Context) (err error) {
 	messageID, err := messageIDFromContext(c)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, webapi.NewErrorResponse(err))
+		return c.JSON(http.StatusBadRequest, jsonmodels.NewErrorResponse(err))
 	}
 
 	if messagelayer.Tangle().Storage.MessageMetadata(messageID).Consume(func(messageMetadata *tangle.MessageMetadata) {
@@ -74,7 +74,7 @@ func GetMessageMetadata(c echo.Context) (err error) {
 		return
 	}
 
-	return c.JSON(http.StatusNotFound, webapi.NewErrorResponse(fmt.Errorf("failed to load MessageMetadata with %s", messageID)))
+	return c.JSON(http.StatusNotFound, jsonmodels.NewErrorResponse(fmt.Errorf("failed to load MessageMetadata with %s", messageID)))
 }
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -86,17 +86,17 @@ func PostPayload(c echo.Context) error {
 	var request PostPayloadRequest
 	if err := c.Bind(&request); err != nil {
 		Plugin().LogInfo(err.Error())
-		return c.JSON(http.StatusBadRequest, webapi.ErrorResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.NewErrorResponse(err))
 	}
 
 	parsedPayload, _, err := payload.FromBytes(request.Payload)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, webapi.ErrorResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.NewErrorResponse(err))
 	}
 
 	msg, err := messagelayer.Tangle().IssuePayload(parsedPayload)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, webapi.ErrorResponse{Error: err.Error()})
+		return c.JSON(http.StatusBadRequest, jsonmodels.NewErrorResponse(err))
 	}
 
 	return c.JSON(http.StatusOK, PostPayloadResponse{ID: msg.ID().String()})
